optimize rotary_emb and move multinomial_sampling into default folder
