+++ lmdeploy/pytorch/kernels/default/__init__.py
@@ -1,1 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .multinomial_sampling import multinomial_sampling
+
+__all__ = [
+    'multinomial_sampling',
+]
