+++ lmdeploy/pytorch/kernels/dipu/apply_rotary_pos_emb.py
@@ -1,6 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import deeplink_ext.cpp_extensions as ext
-import torch
 from torch import Tensor
 
 
@@ -17,13 +16,11 @@ def apply_rotary_pos_emb(
 ):
     bs, head, dim = query_states.shape
     numKeyValueHeads = key_states.shape[1]
-    seqlens = [(min(position_id), max(position_id) + 1)
-               for position_id in position_ids.tolist()]
     query_states = query_states.reshape(bs, head * dim)
     key_states = key_states.reshape(bs, numKeyValueHeads * dim)
     if not (hasattr(context, 'cos') or hasattr(context, 'sin')):
-        cos = torch.cat([cos[i:j] for i, j in seqlens]).view(1, bs, 1, -1)
-        sin = torch.cat([sin[i:j] for i, j in seqlens]).view(1, bs, 1, -1)
+        cos = cos[position_ids_1d].view(1, bs, 1, -1)
+        sin = sin[position_ids_1d].view(1, bs, 1, -1)
         setattr(context, 'cos', cos)
         setattr(context, 'sin', sin)
     ext.rotary_embedding_v2(query_states, key_states, context.cos, context.sin,
